+++ src/utils/isAllDifferent.js
@@ -1,0 +1,5 @@
+const isAllDifferent = (input) => {
+  const unique = new Set(input);
+  return unique.size === input.length ? true : false;
+};
+module.exports = isAllDifferent;
