+++ src/utils/isNumber.js
@@ -1,0 +1,4 @@
+const isNumber = (input) => {
+  return Number.isInteger(Number(input)) ? true : false;
+};
+module.exports = isNumber;
