+++ src/utils/isThree.js
@@ -1,0 +1,4 @@
+const isThree = (input) => {
+  return input.length === 3 ? true : false;
+};
+module.exports = isThree;
